Update params for InputHex and InputHexShort From int to uint64_t and uint32_t respectively

+++ src/app/gui/input.h
@@ -15,8 +15,8 @@ namespace gui {
 constexpr ImVec2 kDefaultModalSize = ImVec2(200, 0);
 constexpr ImVec2 kZeroPos = ImVec2(0, 0);
 
-IMGUI_API bool InputHex(const char* label, int* data);
-IMGUI_API bool InputHexShort(const char* label, int* data);
+IMGUI_API bool InputHex(const char* label, uint64_t* data);
+IMGUI_API bool InputHexShort(const char* label, uint32_t* data);
 
 using ItemLabelFlags = enum ItemLabelFlag {
   Left = 1u << 0u,
